feat: Update reservation status based on date
This commit updates the `DatabaseAdapter` class to include a new method `updateReservationStatus` that updates the status of reservations based on the current date. It also adds two SQL queries to update the reservation status for passed and completed reservations. This change ensures that reservations with a date in the past are marked as "Passed" or "Completed" depending on their status.

Co-authored-by: Mehmet Kuzucu <mehmet-kuzucu@users.noreply.github.com>
Co-authored-by: SevvalCetin <99796036+SevvalCetin@users.noreply.github.com>